Up to C39 working on pedantry mode now.

+++ chapter-27-cellular-automata/src/map_builders/cellular_automata.rs
@@ -112,8 +112,8 @@ impl CellularAutomataBuilder {
         self.take_snapshot();
 
         // Find all tiles we can reach from the starting point
-        let map_starts : Vec<i32> = vec![start_idx as i32];
-        let dijkstra_map = rltk::DijkstraMap::new(self.map.width, self.map.height, &map_starts , &self.map, 200.0);
+        let map_starts : Vec<usize> = vec![start_idx];
+        let dijkstra_map = rltk::DijkstraMap::new(self.map.width as usize, self.map.height as usize, &map_starts , &self.map, 200.0);
         let mut exit_tile = (0, 0.0f32);
         for (i, tile) in self.map.tiles.iter_mut().enumerate() {
             if *tile == TileType::Floor {
