+++ chapter-60-caverns3/src/ai/flee_ai_system.rs
@@ -27,7 +27,7 @@ impl<'a> System<'a> for FleeAI {
             turn_done.push(entity);
             let my_idx = map.xy_idx(pos.x, pos.y);
                 map.populate_blocked();
-                let flee_map = rltk::DijkstraMap::new(map.width, map.height, &flee.indices, &*map, 100.0);
+                let flee_map = rltk::DijkstraMap::new(map.width as usize, map.height as usize, &flee.indices, &*map, 100.0);
                 let flee_target = rltk::DijkstraMap::find_highest_exit(&flee_map, my_idx as i32, &*map);
                 if let Some(flee_target) = flee_target {
                     if !map.blocked[flee_target as usize] {
